remove file if exist

+++ Makefile
@@ -46,7 +46,7 @@ update-dependencies:
 	${COMPOSER} require symfony/asset:~${SYMFONY_VERSION} --no-scripts --no-update
 ifeq ($(SYLIUS_VERSION)$(SYMFONY_VERSION), 2.0.06.4)
 	${COMPOSER} update --no-progress -n --no-scripts
-	rm ${TEST_DIRECTORY}/config/packages/csrf.yaml
+	test -f ${TEST_DIRECTORY}/config/packages/csrf.yaml && rm ${TEST_DIRECTORY}/config/packages/csrf.yaml || true
 endif
 	${COMPOSER} update --no-progress -n
